Remove map_overlap 0-depth workaround
Upstream has since incorporated a change to ensure setting the depth to
`0` in `map_overlap` does not cause an error (for cases other than
`reflect`). As we are bumping the Dask version, we no longer need this
workaround. Hence this drops it to cleanup some code and simplify
maintenance. Also updates the corresponding test to stop checking that
this this workaround is used.