Force dtype of `prod` in `_ravel_shape_indices`
Make sure that NumPy's `prod` returns results that match the `dtype` of
the `arange`. This is important as `prod` by default matches `float` if
no other type is specified (e.g. it is given an empty sequence). While
the `float` value will be handled correctly by NumPy, it's better to
be explicit about the type `prod` should provide. That way if NumPy
changes its behavior, we don't run into issues here.