Use Python 3's definition of division in Python 2 (#58)
Go ahead and switch to using Python 3's definition of division on Python
2 as we are doing division in modules. Though it's worth noting that one
of the values is already floating precision (possibly even complex) when
doing regular division, so this doesn't really affect the result, but it
does make it easier to reason about division in this module. In all
other cases, we are intentionally choosing C-style division in such a
way that it works on Python 2/3 anyways.