Divide `sigma` by `-2` (#59)
Instead of negating the result of `(sigma ** 2) / 2`, which may be a
Dask Array meaning this would add more nodes to the graph, simply divide
by `-2` instead. This ensures the sign still changes as intended, but
this is included as part of the division step. Thus this avoids
potentially adding more nodes to the Dask graph (assuming `sigma` is not
concrete).