Use the unravel indices utility function
Instead of handling the unraveling of a Dask Array of indices through
mathematical operations in Dask (which creates a more complex graph),
Simply use the utility function, which simply uses `map_blocks` on each
chunk to get the N-D indices simply. This makes it easier for Dask to
potentially fuse this task with other tasks. Also keeps the Dask graph
pretty clean.