Use custom kernel for extrema (#61)
* Add extrema kernel

Implements the `_extrema` function to act as a kernel for the `extrema`
function. Effectively fuses all minimum, maximum, and position
computations into the same computational step. Should make this
significantly faster for Dask to compute and the resulting graph should
be much simpler.

* Use `_extrema` kernel

Rewrite the `extrema` function to make use of the new `_extrema` kernel.
This results in a few changes and requires the `extrema` function to
take on some work that the other functions were doing for us. Though the
performance improvements to `extrema` should be worth it.