Update mean's variables to match other functions
As the input image has been summed over, it doesn't really make sense to
include that in the variable names in `mean`. This same rational is used
in other functions. So update `mean`'s variable names accordingly.

Also fix the `com_lbl` name for the result variable in `mean` in the
process. This was an old copy-paste error from the `center_of_mass`
function, which makes no sense in the context of `mean`. It is now
`mean_lbl`, which is more fitting.