Rewrite `histogram` to use `labeled_comprehension`
To simplify the `histogram` implementation a bit, rewrite it using
`labeled_comprehension`. Make use of `functools.partial` to bind
arguments to `histogram` so that it can be used with
`labeled_comprehension`. Also simplify the internal `_histogram`
function into a basic kernel function following these changes. Set the
return type of `labeled_comprehension` as `object` to match that of what
`histogram` used prior.