Test labeled_comprehension with object type (#76)
Includes a test using `labeled_comprehension` with functions that return
`object` type results. While this is also being used by `histogram`
currently, it makes sense to have a test of `labeled_comprehension`
for this case as well in the event that `histogram`'s implementation
changes. The ability to return `object` type results is very handy for
using a variety of functions with `labeled_comprehension`. So it is good
to be sure that `labeled_comprehension` retains the behavior regardless
of other changes.