Add a custom kernel for `center_of_mass`
Create a custom kernel of `center_of_mass` to use with
`labeled_comprehension`. In the kernel take the ROI's intensity and 1-D
positions along with a few arguments to be bound before calling
`labeled_comprehension` (like the shape of the original image and the
results expected type). Use these to unravel the 1-D positions back into
N-D positions and computed the center of mass for the given ROI to pass
back in an expected form. As `labeled_comprehension` can only handle
scalars being returned for each ROI, use a structured type to pass back
each coordinate of the center of mass in a different field. The
structured type uses a simple naming scheme based on a string holding
the index corresponding to that dimension.