Use N-D structured array in `minimum_position`
Rewrite `minimum_position` to use a structured `dtype` with a 1-D
field.  This allows the unraveled position to be determined in the
`minimum_position` kernel function and returned. With this change there
is no need to handle unraveling of the position in the Dask graph. Only
the structured field need be selected, which is a constant time
operation. Thus simplifying the Dask graph generated by
`minimum_position` and improving the computation performance as well.