Cleanup `test_labeled_comprehension_struct` (#80)
* Simplify `test_labeled_comprehension_struct` case

For some reason this was using `argmax` and then selecting the max value
from the array. In reality, this can just find the max value and return
it. So this simplifies the test accordingly.

* Drop extra blank line

As both of these are being defined to fill out values in
`labeled_comprehension`, go ahead and group them together.