Construct mask for each label right before using
In `labeled_comprehension`, instead of creating a larger array of masks
for each index before running the user defined function, create each
mask for each label right before using it to select out relevant data
for the user defined function. This has the benefit that computing any
result for any label is not necessary blocked by the computation of any
other label. Dask can usually optimize such a graph better as well. Not
to mention the Dask graph is a bit simpler because of this change
anyways. Improves the performance and cleans up the code too.