Use NumPy's `in1d` instead of `isin`
As older versions of NumPy that we support and test against don't
include `isin`, switching to using `in1d`, which has been around longer.
Since the array in question is already 1-D, there is no need for us to
reshape the result after calling `in1d`. So this is sufficient for our
use case.