Handle empty adjacency graph for singleton chunk
When there is a singleton chunk, there are no shared faces between
chunks to construct an adjacency graph from. In this case ensure there
is at least an empty array to start with. This doesn't alter the cases
where there are multiple chunks that do share faces. Though it does
avoid branching when there is a single chunk with no shared faces.