Ensure `n` is a `LABEL_DTYPE` scalar
To make sure that `n` matches the type we want it to be, add a delayed
cast to the desired type. Normally `n` would be of type `int`, but we
need it to `LABEL_DTYPE` so we can add it to the label images without
causing a type conversion. So this fixes that issue as well.