Implement `_unique_axis_0` by with structured view
By viewing the array as a structured array that merely groups together
the other dimensions within the structure, `_unique_axis_0` is able to
call NumPy's `unique` function on the array keeping the type information
unchanged. Thus if `unique` is able to handle the specific type more
efficiently, we benefit from that speed up still.

Note: More work would be needed if we wanted to handle F-contiguous
arrays, but that is not needed for our use case.