Optimize out first `where` in `label` (#102)
* Store index and slice iterator before `for`-loop

* Handle `input` block selection in iterator

* Optimize out first `where` in `label`

* Move input selection within `zip`

As `starmap` is less familiar to others, move the selection operation
within `zip` using `map`. This should also make it clearer what the
selection is being used for and what is being selected from.

* Use Python 2/3 compatibility functions

Make use of iterator friendly Python 2/3 `imap` and `izip` functions
when generating the `block_iter`. This should ensure things like `next`
behave correctly (particularly on Python 2).

* Replace `lambda` with `operator.getitem`

Uses `operator.getitem` to make it a bit clearer to the reader as to
what is happening in this map operation (as opposed to the `lambda`).