Dask impl. of skimage.exposure.match_histograms
Same structure and testing as original, but tests include chunking that would fail a naive (simply parallel) implmentation.

The non-trivial need is an interpolation across unique values.  Until np.interp has a dask implementation, a bottleneck occurs in which the full range of values might be held in memory.  This is feasible for dtypes up to int16, but likely should be avoided for int32 and higher, or float types.