Update dependencies in RTD environment (#168)
* Update dependencies in RTD environment

ReadTheDocs builds recently started failing. This seems to be due to a
`pip` option being passed that the version of `pip` we use here doesn't
understand. In an attempt to fix this issue, try modernizing our
dependencies here.

https://readthedocs.org/projects/dask-image/builds/12159489/

* Bump `pip` and `wheel` to latest

Apparently `pip` version `20.2`+ has the newer resolver, which RTD tries
to use. So bump to that version so that builds don't fail due to this
flag being passed.