Fix wrap bug in ndfilters convolve and correlate (#243)
* Bug fixing ndfilters.convolve and ndfilter.correlate

* Added tests

* Fixed a syntax error

* Update dask_image/ndfilters/_conv.py

Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>

* Update dask_image/ndfilters/_conv.py

Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>

Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>

Why do we set `mode=constant` here? "The reason I did it is because I assume that it is (barely) faster than doing reflect because padding with constant values is easier than padding reflectively." -  anlavandier
Gen also assumes that it is unlikely to cause edge artifacts between chunks, because the user can adjust the map_overlap depth to the point where the scipy chunk boundary modes don't really affect the final Dask result.